TRAN-511 Add UI improvements and fixes

+++ app/src/main/java/mega/privacy/android/app/presentation/extensions/transfers/InProgressTransferExt.kt
@@ -68,7 +68,7 @@ internal fun InProgressTransfer.getProgressString(context: Context, isOverQuota:
         val totalSizeString = getTotalSizeString(context)
         val progressString = getString(
             R.string.progress_size_indicator,
-            progress.floatValue.toInt(),
+            progress.intValue,
             totalSizeString
         )
